NodeDefView: moved components into webapp/views/App/views/NodeDef - introduced NodeDefView local state
