+++ webapp/views/App/views/Designer/Designer.js
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux'
 
 import SurveyDefsLoader from '@webapp/components/survey/SurveyDefsLoader'
 import ModuleSwitch from '@webapp/components/moduleSwitch'
-import NodeDefView from '@webapp/loggedin/surveyViews/nodeDef/nodeDefView'
+import NodeDefView from '@webapp/views/App/views/NodeDef'
 import CategoriesView from '@webapp/loggedin/surveyViews/categories/categoriesView'
 import CategoryView from '@webapp/loggedin/surveyViews/category/categoryView'
 import TaxonomiesView from '@webapp/loggedin/surveyViews/taxonomies/taxonomiesView'
