+++ webapp/views/App/views/NodeDef/advanced/AdvancedProps.js
@@ -9,7 +9,7 @@ import { useI18n } from '@webapp/store/system'
 import { FormItem } from '@webapp/components/form/input'
 import Checkbox from '@webapp/components/form/checkbox'
 
-import NodeDefExpressionsProp from './expressionsProp/nodeDefExpressionsProp'
+import NodeDefExpressionsProp from './ExpressionsProp/NodeDefExpressionsProp'
 
 const AdvancedProps = (props) => {
   const { nodeDef, validation, nodeDefParent, setNodeDefProp, readOnly } = props
