+++ webapp/views/App/views/NodeDef/advanced/ExpressionsProp/NodeDefExpressionsProp.js
@@ -6,7 +6,7 @@ import * as NodeDefExpression from '@core/survey/nodeDefExpression'
 import * as Validation from '@core/validation/validation'
 import * as Expression from '@core/expressionParser/expression'
 
-import ExpressionsProp from './expressionsProp'
+import ExpressionsProp from './ExpressionsProp'
 
 const NodeDefExpressionsProp = props => {
   const {
