+++ webapp/views/App/views/NodeDef/index.js
@@ -1,0 +1,1 @@
+export { default } from './NodeDef'
